==> R/00_all.qmd <==
---
title: "Full Analysis Pipeline"
format: html
editor: visual
execute:
  cache: false
---

```{r}
library("tidyverse")
```

```{r}
#| child:
#|   - "01_load.qmd"
#|   - "02_clean.qmd"
#|   - "03_augment.qmd"
#|   - "04_describe.qmd"
#|   - "05_analysis1.qmd"
#|   - "06_analysis2.qmd"
```
