==> R/02_clean.qmd <==
---
title: "02 - clean"
format: html
editor: visual
execute:
  cache: false
---

# Load libraries

```{r}
library("tidyverse")
library("dplyr")
```

# Clean gene table

```{r}
gen_exp <- readRDS("../data/gen_exp.rds")
dim(gen_exp)
gen_exp |> 
  slice_sample(n = 10)
```

We have 60600 observations (rows) and 430 variables (columns). The rows are the different genes and the columns are the samples. We have 60600 genes (that is too much) and we can see that some of them are 0 for all the samples. So we want to delete all the gene with no expression for each sample.

```{r}
gen_exp_clean <- gen_exp |>
  filter(rowSums(across(-Ensembl_ID)) != 0)
dim(gen_exp_clean)
```

We still have too much genes (57825) so we want to see which genes are useless

We calculate the sum of all the gene expression for each gene

```{r}
gen_exp_clean <- gen_exp_clean |>
  mutate(sum_gene_exp = rowSums(across(-Ensembl_ID))) |>
  relocate(sum_gene_exp, .after = Ensembl_ID) |>
  arrange(sum_gene_exp)
gen_exp_clean |> 
  slice_sample(n = 10)
```

Now we want to visualize this new variable to know what threshold we are going to use to select the relevant genes

```{r}
ggplot(gen_exp_clean, aes(x = Ensembl_ID, y = sum_gene_exp)) +
  geom_point()
ggplot(gen_exp_clean, aes(x = sum_gene_exp)) +
  geom_histogram(binwidth = 100)
ggplot(gen_exp_clean, aes(y = sum_gene_exp)) +
  geom_boxplot()
```

We can see that a lot of genes have the sum around 0 (with the boxplot). In the histogram, the binwidth is 100 and 4 bars are higher than the others and correspond to irrelevant genes. So we can choose a threshold of 400. There are 431 samples so a threshold of 400 means that in average, approximately all the genes expressions are higher than 1.

```{r}
gen_exp_clean <- gen_exp_clean |>
  filter(sum_gene_exp > 400)
dim(gen_exp_clean)
```

```{r}
ggplot(gen_exp_clean, aes(x = sum_gene_exp)) +
  geom_histogram(binwidth = 100)
```

Now, we have 14157 genes which is more appropriate for our case.

```{r}
gen_exp_clean <- gen_exp_clean |>
  select(!c(sum_gene_exp))
```

```{r}
saveRDS(gen_exp_clean, "../data/gen_exp_clean.rds")
```

# Cleaning Phenotype table

## Loading table

We start by loading the "phenotype" table, with the raw clinical data. The slice_sample allows to see 20 random samples and start evaluating which variables (columns) will be useful

```{r}
phenotype <- readRDS("../data/phenotype.rds")
phenotype |> 
  slice_sample(n = 20)
```

### Removal of variables

We start by removing some unuseful variables, such as:

-   "id", "case_id", "tissue_source_site_id.tissue_source_site", "entity_id.annotations", "annotation_id.annotations" and other:\
    They are long lists of seemingly random characters, useful for identification of the sample but not for analysis, as we have no information on their meaning.\
    We'll use "sample" as identification.

-   "submitter_id", "entity_submitter_id.annotations" :\
    They are the same as "sample" without the last 3 characters. This is repetitive and untidy.

```{r}
# Shows how some variables repeat the same information 
phenotype |> 
  select(c(sample, submitter_id)) |> 
  slice_sample(n = 20) 

# Shows identifier variables, no meaning for analysis 
phenotype |> 
  select(c(id, case_id)) |> 
  slice_sample(n = 20) 

# Remove these variables 
phenotype_clean <- phenotype |> 
  select(!c(id, 
            case_id, 
            submitter_id, 
            tissue_source_site_id.tissue_source_site, 
            entity_submitter_id.annotations, 
            entity_id.annotations, 
            annotation_id.annotations, 
            case_id.annotations, 
            treatment_id.treatments.diagnoses, 
            submitter_id.treatments.diagnoses, 
            sample_id.samples, 
            pathology_report_uuid.samples)
         )

  
```

-   "primary_site", "primary_site.project", "project_id.project", ... : These columns contain the same observation for each row, and will not be useful for analysis either.\
    For example "gender.demographic" will always be "female", as we are studying ovarian cancer.

```{r}
# Shows variables with repeated observations 
phenotype_clean |> 
  select(c(primary_site, 
           name.program.project, 
           last_known_disease_status.diagnoses, 
           gender.demographic)) |> 
  slice_sample(n = 20) 

# Remove these variables 
phenotype_clean <- phenotype_clean |> 
  select(!c(primary_site, 
            primary_site.project, 
            project_id.project, 
            disease_type.project, 
            name.project, 
            name.program.project, 
            project.tissue_source_site, 
            days_to_diagnosis.diagnoses, 
            last_known_disease_status.diagnoses, 
            tissue_or_organ_of_origin.diagnoses, 
            preservation_method.samples, 
            is_ffpe.samples, 
            gender.demographic)
         )
```

-   "entity_type.annotations" , "state.annotations", ... : These are empty or always the same when not empty. Samples might come from different sources, which would explain irregular conventions. After careful reading of the table, we can say that the fact that some observations are left empty has no other meaning.

```{r}
# Shows variables with repeated or empty observations 
phenotype_clean |> 
  select(c(entity_type.annotations, 
           state.annotations, 
           status.annotations)) |> 
  slice_sample(n = 20) 

# Remove these columns 
phenotype_clean <- phenotype_clean |> 
  select(!c(entity_type.annotations, 
            state.annotations, 
            status.annotations, 
            composition.samples, 
            days_to_sample_procurement.samples, 
            state.treatments.diagnoses,
            tumor_grade.diagnoses,
            progression_or_recurrence.diagnoses))

```

The samples with a disease_type value other than "Cystic, Mucinous and Serous Neoplasms" miss a lot of other information (NA in other columns). We will not keep them as there are only 8 samples and will disturb analysis without providing new information. It now makes no sense to keep this column.

```{r}
# Check for other disease types 
phenotype_clean |> 
  filter(disease_type != "Cystic, Mucinous and Serous Neoplasms") 

# Remove samples with missing values 
phenotype_clean <- phenotype_clean |> 
  filter(disease_type == "Cystic, Mucinous and Serous Neoplasms")

# Remove this column
phenotype_clean <- phenotype_clean |> 
  select(!disease_type)
```

The column prior_treatment.diagnoses is Not Reported or No for every samples except one Yes. This column will not be useful.

```{r}
# Show values other than "Not Reported" and "No" for prior_treatment.diagnoses
phenotype_clean |> 
  filter(prior_treatment.diagnoses!="Not Reported" &
         prior_treatment.diagnoses != "No") |> 
  relocate(prior_treatment.diagnoses, .before = 1)
```

Some variables don't always contain the same information on paper, but further analysis shows us that the observations are very similar. For example alcohol_history.exposures is "Not reported" or NA:

```{r}
# Show NAs in alcohol_history.exposures
phenotype_clean |> 
  filter(is.na(alcohol_history.exposures)) 

# Remove this column
phenotype_clean <- phenotype_clean |> 
  select(!alcohol_history.exposures) 

```

Modification of the variable vital_status.demographic to a binary variable "alive"

```{r}
phenotype_clean <- phenotype_clean |> 
  mutate(vital_status.demographic = case_when(
    vital_status.demographic == "Alive" ~ TRUE,
    vital_status.demographic == "Dead" ~ FALSE,
    TRUE ~ NA
  ))|>
  rename(alive=vital_status.demographic)
```

```{r}
saveRDS(phenotype_clean, "../data/phenotype_clean.rds")

```
